backend/requirements.txt
Django<2.3,>=2.2
djangorestframework
djoser
python-dotenv
pytest
pytest-django
django-stubs
djangorestframework-stubs
flake8
flake8-annotations
flake8-broken-line
flake8-docstrings
flake8-return
flake8-isort
